+++ fe/src/App.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import VendingMachine from 'pages/VendingMachine';
+import VMmangement from 'pages/VMmangement';
 import Layout from './layout';
 
 function App() {
@@ -9,6 +10,7 @@ function App() {
       <Routes>
         <Route path="/" element={<Layout />}>
           <Route index element={<VendingMachine />} />
+          <Route path="mangement" element={<VMmangement />} />
         </Route>
       </Routes>
     </Router>
