+++ fe/src/components/GlobalNavigation/style.js
@@ -1,0 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+import styled from 'styled-components';
+
+export const Header = styled.header`
+  width: 100%;
+  height: 200px;
+`;
