+++ fe/src/layout/style.js
@@ -4,11 +4,6 @@ export const AppContainer = styled.div`
   width: 100vw;
 `;
 
-export const Header = styled.header`
-  width: 100%;
-  height: 200px;
-`;
-
 export const Main = styled.main`
   width: 1280px;
   margin: 0 auto;
