+++ fe/src/pages/VMmangement/style.js
@@ -1,0 +1,13 @@
+/* eslint-disable import/prefer-default-export */
+import styled from 'styled-components';
+
+export const ChangesUnitsContainer = styled.ul`
+  width: 20%;
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  > li {
+    display: flex;
+    gap: 12px;
+  }
+`;
